+++ convex/http.ts
@@ -1,6 +1,6 @@
 import { httpRouter } from 'convex/server';
 import { auth } from './auth';
-import { polarWebhook } from './topUps/public';
+import { handlePolarWebhook } from './utils/polar';
 
 const http = httpRouter();
 
@@ -9,7 +9,7 @@ auth.addHttpRoutes(http);
 http.route({
 	path: '/polar/webhook',
 	method: 'POST',
-	handler: polarWebhook,
+	handler: handlePolarWebhook,
 });
 
 export default http;
